Fix codelens document selector and decl refresh
