+++ src/diagram/generator.ts
@@ -1,11 +1,11 @@
 import { SGraph, SEdge, SNode, SPort, Point, SLabel, Dimension } from 'sprotty-protocol';
 import { ElkExtendedEdge, ElkGraphElement, ElkLabel, ElkNode, ElkPort } from 'elkjs/lib/elk.bundled.js';
-import { DeclCollector, SymbolType } from "../decl";
+import { DeclCollector, SymbolType } from "../passes/decl";
 import { ComponentDecl, ComponentInstanceDecl, Connection, DirectGraphDecl, GeneralInputPortInstance, GeneralPortKind, IncludeStmt, InterfaceImportStmt, IntExprValue, PortInstanceDecl, PrimExprType, QualifiedIdentifier, SpecialOutputPortInstance, SpecialPortKind, TopologyDecl } from "../parser/ast";
-import { MemberTraverser } from "../traverser";
+import { MemberTraverser } from "../passes/traverser";
 import type { ComponentSNode, PortSNode } from '../../common/models';
 import { getInterfaceFullnameFromImport } from '../util';
-import { FppAnnotator } from '../annotator';
+import { FppAnnotator } from '../passes/annotator';
 import { ExprTraverser } from '../evaluator';
 
 /**
