+++ src/passes/decl.ts
@@ -1,10 +1,10 @@
 import * as vscode from 'vscode';
-import * as Fpp from './parser/ast';
+import * as Fpp from '../parser/ast';
 
 import { MemberTraverser } from './traverser';
-import { ReferenceTracker } from './referenceTracker';
-import { RangeAssociator } from './associator';
-import { DiangosicManager } from './diagnostics';
+import { ReferenceTracker } from '../referenceTracker';
+import { RangeAssociator } from '../associator';
+import { DiangosicManager } from '../diagnostics';
 
 export enum SymbolType {
     module,
@@ -130,133 +130,33 @@ export class DeclCollector extends MemberTraverser {
     // This way we can clean and refresh a single file's declarations
     translationUnitDeclarations = new Map<string, RangeAssociator<[SymbolType, string]>>();
 
-    // If a translation unit includes component instances,
-    // there is a second stage of reference depth that should
-    // force re-annotation on doc refresh. Normally document
-    // annotation refresh happens only on open text documents
-    // since its not required, in this case it shouldn't matter
-    // if the doc is open or not
-    hasComponentInstances: boolean = false;
-
-    topologyPortsTrav = new class extends MemberTraverser {
-        parent!: DeclCollector;
-
-        private update(ast: Fpp.ComponentInstanceDecl, componentDef: Fpp.ComponentDecl, member: Fpp.ComponentMember, componentScope: Fpp.QualifiedIdentifier) {
-            switch (member.type) {
-                case 'GeneralPortInstanceDecl': {
-                    const portName = MemberTraverser.flat(componentScope, member.name);
-                    if (member.kind.isOutput) {
-                        this.parent.outputPortInstances.set(portName, member);
-                        this.parent.translationUnitDeclarations.get(ast.location.source)!.add(
-                            DiangosicManager.asRange(ast.name.location),
-                            [SymbolType.outputPortInstance, portName]
-                        );
-                    } else {
-                        this.parent.inputPortInstances.set(portName, member);
-                        this.parent.translationUnitDeclarations.get(ast.location.source)!.add(
-                            DiangosicManager.asRange(ast.name.location),
-                            [SymbolType.inputPortInstance, portName]
-                        );
-                    }
-                    break;
-                }
-                case 'SpecialPortInstanceDecl': {
-                    const portName = MemberTraverser.flat(componentScope, member.name);
-                    switch (member.kind.kind.value) {
-                        case 'commandRecv':
-                        case 'commandReg':
-                        case 'commandResp':
-                        case 'event':
-                        case 'paramGet':
-                        case 'paramSet':
-                        case 'telemetry':
-                        case 'textEvent':
-                        case 'timeGet':
-                            // Auto-ports
-                            break;
-                        case 'productGet':
-                        case 'productRequest':
-                        case 'productSend':
-                            this.parent.outputPortInstances.set(portName, member);
-                            this.parent.translationUnitDeclarations.get(ast.location.source)!.add(
-                                DiangosicManager.asRange(ast.name.location),
-                                [SymbolType.outputPortInstance, portName]
-                            );
-                            break;
-                        case 'productRecv':
-                            this.parent.inputPortInstances.set(portName, member);
-                            this.parent.translationUnitDeclarations.get(ast.location.source)!.add(
-                                DiangosicManager.asRange(ast.name.location),
-                                [SymbolType.inputPortInstance, portName]
-                            );
-                    }
-                    break;
-                }
-                case 'IncludeStmt':
-                    if (member.resolved) {
-                        for (const m of member.resolved.members) {
-                            this.update(ast, componentDef, m, componentScope);
-                        }
-                    }
-                    break;
-                // If this component imports port interfaces,
-                // add the imported ports to the port instance map.
-                case 'InterfaceImportStmt':
-                    // Get the interface decl from combining the def scope and the interface name.
-                    // If member.symbol has more than one identifier (e.g., `import Drv.Tick`), that means
-                    // the interface is defined outside the current scope. In this case, do not prepend
-                    // the component def's scope. But if member.symbol has only one identifier (e.g., `import Framer`),
-                    // that means the interface is defined under the same scope as the component def.
-                    // In this case, prepend the component def's scope (e.g., Svc). Note that only
-                    // the first identifier is prepended.
-                    let interfaceFullName = undefined;
-                    if (member.symbol.length > 1) {
-                        interfaceFullName = MemberTraverser.flat(member.symbol);
-                    } else if (member.symbol.length === 1) {
-                        interfaceFullName = `${componentDef.scope[0].value}.${MemberTraverser.flat(member.symbol)}`;
-                    } else {
-                        console.error("Interface full name undefined.");
-                        return;
-                    }
-
-                    // Get the interface decl.
-                    const interfaceDecl = this.parent.interfaces.get(interfaceFullName);
-                    if (!interfaceDecl) {
-                        console.error(`interfaceDecl for ${interfaceFullName} not found`);
-                        return;
-                    }
-
-                    // Update all ports from the interface decl.
-                    interfaceDecl.members.map(m => {
-                        this.update(ast, componentDef, m, componentScope);
-                    });
-                    break;
-            }
-        }
-
-        protected componentInstanceDecl(ast: Fpp.ComponentInstanceDecl, scope: Fpp.QualifiedIdentifier): void {
-            const resolved = this.parent.resolve(ast.fppType.type, scope, SymbolType.component);
-            if (!resolved) {
-                // This error will be annotated later
-                return;
-            }
-
-            this.parent.hasComponentInstances = true;
-
-            const componentScope = [...scope, ast.name];
-
-            const componentAst = this.parent.components.get(MemberTraverser.flat(resolved));
-            if (componentAst) {
-                for (const member of componentAst.members) {
-                    this.update(ast, componentAst, member, componentScope);
-                }
-            }
-        }
-    };
-
-    constructor() {
-        super();
-        this.topologyPortsTrav.parent = this;
+    clearAll() {
+        this.actions.clear();
+        this.commands.clear();
+        this.components.clear();
+        this.componentInstances.clear();
+        this.constants.clear();
+        this.containers.clear();
+        this.dictionary.clear();
+        this.events.clear();
+        this.generalInputPortDecl.clear();
+        this.generalOutputPortDecl.clear();
+        this.graphGroups.clear();
+        this.guards.clear();
+        this.inputPortInstances.clear();
+        this.interfaces.clear();
+        this.outputPortInstances.clear();
+        this.parameters.clear();
+        this.ports.clear();
+        this.records.clear();
+        this.references = new ReferenceTracker<vscode.Range>();
+        this.signals.clear();
+        this.states.clear();
+        this.stateMachines.clear();
+        this.stateMachineInstances.clear();
+        this.telemetry.clear();
+        this.topologies.clear();
+        this.types.clear();
     }
 
     clearDecls(grammarSource: string) {
@@ -354,14 +254,12 @@ export class DeclCollector extends MemberTraverser {
     }
 
     pass(ast: Fpp.TranslationUnit, scope: Fpp.QualifiedIdentifier = []): void {
-        this.hasComponentInstances = false;
         this.clearDecls(ast.location.source);
         for (const _ of ast.dependencies) {
             this.clearDecls(ast.location.source);
         }
 
         super.pass(ast, scope);
-        this.topologyPortsTrav.pass(ast, scope);
     }
 
     static typeName(type: Fpp.TypeName): string {
